+++ edb/edgeql-rust/src/lib.rs
@@ -1,22 +1,23 @@
 #[macro_use]
 extern crate cpython;
 
-use cpython::PyString;
-
 mod errors;
 mod float;
 mod hash;
 mod keywords;
 pub mod normalize;
+mod parser;
 mod position;
 mod pynormalize;
 mod tokenizer;
 
-use edgeql_parser::{into_python::IntoPython, parser};
-use errors::TokenizerError;
-use position::{offset_of_line, SourcePoint};
-use pynormalize::normalize;
-use tokenizer::{get_unpickle_fn, tokenize, Token};
+use cpython::PyString;
+
+use crate::errors::TokenizerError;
+use crate::parser::{parse_block, parse_single as parse_sing};
+use crate::position::{offset_of_line, SourcePoint};
+use crate::pynormalize::normalize;
+use crate::tokenizer::{get_unpickle_fn, tokenize, Token};
 
 py_module_initializer!(
     _edgeql_rust,
@@ -48,17 +49,8 @@ py_module_initializer!(
         m.add(py, "partial_reserved_keywords", keywords.partial)?;
         m.add(py, "future_reserved_keywords", keywords.future)?;
         m.add(py, "current_reserved_keywords", keywords.current)?;
-        m.add(py, "parse", py_fn!(py, parse(data: &PyString)))?;
+        m.add(py, "parse_block", py_fn!(py, parse_block(data: &PyString)))?;
+        m.add(py, "parse_single", py_fn!(py, parse_sing(data: &PyString)))?;
         Ok(())
     }
 );
-
-fn parse(py: cpython::Python, source: &PyString) -> cpython::PyResult<cpython::PyObject> {
-    let source = source.to_string(py)?;
-
-    // parse
-    let ast_root = parser::parse(&source);
-
-    // convert into a Python object
-    ast_root.into_python(py, None)
-}
